requirements.txt
paho-mqtt==1.5.1
pymodbus==2.4.0
pyyaml==5.3.1